Create function to dynamically draw clocks
This update required editing the previous version of the analog clock
by reducing it's size and changing the html structure and css elements.

This was done to make room for the new clocks drawn at the bottom half
of the page which will show the time for time zones around the world.

To achieve the current layout a new function was made drawCityClocks
which dynamically draws a div container, then two more div's are
created for the city name and its corresponding clock. This makes it
easy to reduce or add the number of clocks being drawn by changing the
maximum number in the loop instead of manually creating or removing
div's in the html file.